+++ packages/vue/src/index.ts
@@ -1,4 +1,6 @@
 export * from './composables/powerSync';
-export * from './composables/usePowerSyncQuery';
-export * from './composables/usePowerSyncStatus';
-export * from './composables/usePowerSyncWatchedQuery';
+export { usePowerSyncQuery } from './composables/usePowerSyncQuery';
+export { usePowerSyncStatus } from './composables/usePowerSyncStatus';
+export { usePowerSyncWatchedQuery } from './composables/usePowerSyncWatchedQuery';
+export { useQuery } from './composables/useQuery';
+export { useStatus } from './composables/useStatus';