+++ packages/kysely-driver/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @powersync/kysely-driver
 
+## 0.2.0
+
+### Minor Changes
+
+- 3e92a9c: Made `destroy` and `releaseConnection` no-op functions. If you relied on `destroy` you will need to use `disconnectAndClear` on the PowerSync DB directly.
+
+### Patch Changes
+
+- Updated dependencies [c94be6a]
+- Updated dependencies [2f1e034]
+- Updated dependencies [21801b9]
+  - @powersync/common@1.7.0
+
 ## 0.1.9
 
 ### Patch Changes