+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @powersync/vue
 
+## 0.1.0
+
+### Minor Changes
+
+- a5550b2: Introduced `useQuery` and `useStatus` composables which include compilable query support. Deprecated `usePowerSyncQuery`, `usePowerSyncWatchedQuery`, and `usePowerSyncStatus`.
+
+### Patch Changes
+
+- Updated dependencies [c94be6a]
+- Updated dependencies [2f1e034]
+- Updated dependencies [21801b9]
+  - @powersync/common@1.7.0
+
 ## 0.0.4
 
 ### Patch Changes