+++ tools/diagnostics-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # diagnostics-app
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [c94be6a]
+- Updated dependencies [d62f367]
+  - @powersync/react@1.3.0
+  - @powersync/web@0.5.2
+
 ## 0.1.1
 
 ### Patch Changes